tabla de frecuencias funcinando (distrib unif)
